Change approach to data transformation
Now we will prepare a micro-batch of data by joining Open Meteo
and PV Output data for a given date range, and do this iteratively
on the existing corpus of CSV files. We can reuse the micro-batch
logic in future to incrementally update the training data set on a
regular basis.